--- src/main/java/Command.java
@@ -1,7 +1,0 @@
-abstract class Command {
-    public abstract void exec(TaskList tasks, Ui ui, Storage storage) throws NoListException;
-
-    public boolean isExit() {
-        return false;
-    }
-}
