+++ src/main/java/Duke.java
@@ -1,8 +1,9 @@
-import java.io.FileNotFoundException;
-import java.io.IOException;
-import java.util.Scanner;
-import java.io.File;
-import java.io.FileWriter;
+import duke.Parser;
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+import duke.command.Command;
+import duke.dukeexception.DukeException;
 
 public class Duke {
