--- src/main/java/EmptyDescriptionException.java
@@ -1,5 +1,0 @@
-class EmptyDescriptionException extends DukeException {
-    public EmptyDescriptionException(String message) {
-        super(message);
-    }
-}
