--- src/main/java/IllegalCommandException.java
@@ -1,5 +1,0 @@
-class IllegalCommandException extends DukeException {
-    public IllegalCommandException(String message) {
-        super(message);
-    }
-}
