--- src/main/java/List.java
@@ -1,5 +1,0 @@
-class List extends Command {
-    public void exec(TaskList tasks, Ui ui, Storage storage) {
-        tasks.display();
-    }
-}
