+++ src/main/java/duke/command/Command.java
@@ -1,0 +1,14 @@
+package duke.command;
+
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+import duke.dukeexception.NoListException;
+
+public abstract class Command {
+    public abstract void exec(TaskList tasks, Ui ui, Storage storage) throws NoListException;
+
+    public boolean isExit() {
+        return false;
+    }
+}
