+++ src/main/java/duke/command/Exit.java
@@ -1,4 +1,11 @@
-class Exit extends Command {
+package duke.command;
+
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+import duke.command.Command;
+
+public class Exit extends Command {
     public void exec(TaskList tasks, Ui ui, Storage storage) {
         System.out.println("Bye. Hope to see you again soon!");
     }
