+++ src/main/java/duke/command/List.java
@@ -1,0 +1,12 @@
+package duke.command;
+
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+import duke.command.Command;
+
+public class List extends Command {
+    public void exec(TaskList tasks, Ui ui, Storage storage) {
+        tasks.display();
+    }
+}
