+++ src/main/java/duke/dukeexception/DukeException.java
@@ -1,4 +1,6 @@
-class DukeException extends Exception {
+package duke.dukeexception;
+
+public class DukeException extends Exception {
     public DukeException(String message) {
         super(message);
     }
