+++ src/main/java/duke/dukeexception/EmptyDescriptionException.java
@@ -1,0 +1,7 @@
+package duke.dukeexception;
+
+public class EmptyDescriptionException extends DukeException {
+    public EmptyDescriptionException(String message) {
+        super(message);
+    }
+}
