+++ src/main/java/duke/dukeexception/IllegalCommandException.java
@@ -1,0 +1,7 @@
+package duke.dukeexception;
+
+public class IllegalCommandException extends DukeException {
+    public IllegalCommandException(String message) {
+        super(message);
+    }
+}
