+++ src/main/java/duke/dukeexception/NoListException.java
@@ -1,7 +1,9 @@
+package duke.dukeexception;
+
 import java.io.FileNotFoundException;
 import java.io.IOException;
 
-class NoListException extends DukeException {
+public class NoListException extends DukeException {
 
     public NoListException(Exception e) {
         super(e);
