+++ src/main/java/duke/task/Deadline.java
@@ -1,10 +1,12 @@
+package duke.task;
+
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.time.format.DateTimeFormatter;
 import java.util.Collections;
 import java.util.Arrays;
 
-class Deadline extends Task {
+public class Deadline extends Task {
 
     protected LocalDate date;
     protected LocalTime time;
