+++ src/main/java/duke/task/Event.java
@@ -1,4 +1,8 @@
-class Event extends Task {
+package duke.task;
+
+import duke.task.Task;
+
+public class Event extends Task {
     protected String at;
 
     public Event(String description, String at) {
