+++ src/main/java/duke/task/Todo.java
@@ -1,4 +1,8 @@
-class Todo extends Task{
+package duke.task;
+
+import duke.task.Task;
+
+public class Todo extends Task {
 
     public Todo(String description) {
         super(description);
